Implement email check before user registration
Implemented an async function in member.service.js to check if the user's email is already registered. If it is, a 400 status response along with an "Email already exist" message is sent back in the register function in member.controller.js file. This check prevents duplicate registrations under the same email.