Add default anime list for new members
In the member registration process, a new step was added to create a default anime list for each new member automatically. This list includes categories of "En cours", "A voir", "Terminé", and "Favoris". Furthermore, the association between "Member" and "AnimeList" in models was updated to include a "not null" foreign key constraint. The "checkEmail" function was also refactored for simplicity in member services.