Implement new anime list and related functionality
Expanded the backend's functionality by creating a new anime list router, controller, validator, and service. The new animeList.route.js is integrated into the main route, and the animeList.controller.js lays out a structure for various operations including adding new anime to the list. The validator and dto for the anime list were also created. Now, a token comes as an object instead of a string in the member.controller.js file.