Add delete functionality to anime list
This commit introduces the delete operation in the anime list. Specifically, code was added in animeList.router.js to delete an anime using the /:id route. Accompanying logic was also crafted on animeList.controller.js to handle deletion which includes error handling and response setting. Lastly, the animeList.service.js was modified to execute the actual removal of the anime from the list.