Implement update functionality in anime list
The update operation was added to the anime list module. The router now has an endpoint for updating an item in the anime list. The corresponding logic in the controller was introduced to validate the data, handle errors, and setup appropriate responses. Back-end service was updated to execute the database query for updating an anime in the list.