Implement functionality for adding anime to list
Added a new route in animeList.router.js to add an anime to a list. Implemented the corresponding logic in animeList.controller.js that uses updated anime.validator.js and anime.dto.js for validating and returning the right data. This functionality handles the validation and properly responds with appropriate HTTP status codes in case of invalid data or unsuccessful addition.