Add GET route to fetch anime by id
Introduced a new GET route on the anime router that allows fetching of individual anime by id. To effectively handle the request, adjustments were made in the anime controller and service to retrieve the anime information from the database. Error handling has been implemented in case an anime with the provided id does not exist.