Implement interactive search with anime service
The SearchBar component has been updated to allow active search of animes. It handles input changes and fetches data from the anime service when the form is submitted. In addition, the anime service now communicates with the Jikan API, returning the queried data. Also, a new state for storing the found anime list was introduced in the Dashboard component.