Implement function to move finished anime
This commit introduces a check for whether an anime series has been watched in full and it's series status has been updated accordingly. The function increase in episode service will identify if an anime has been fully watched and will move it to the appropriate list. This functionality is being used in the episode control and is applied in the frontend through the AnimeInListItem component as well.