Implement IDs tracking for watched anime episodes
This commit introduces tracking of anime IDs for watched episodes in the AnimeInListItem component. The update makes use of the Set data structure to add or delete anime IDs as episodes are watched or updated. Changes are reflected in the incrementEpisodes and decrementEpisodes functions, ensuring only the necessary anime remain on the list.