Refactor UI layout and update backend routes
This commit refactors the frontend UI layouts, particularly in the Header component for better user interaction, and enhances the styling. The Manga Gest anime recommendation system was improved by integrating user preferences with backend server calls, which also led to modification of API routes. Additionally, unnecessary logging was removed and 'synopsis' field size in the Anime model was adjusted for brevity.