Add list title to AnimeInListItem
The list title is now passed as a prop to the AnimeInListItem component, which is then passed to the AnimeItem component. Some changes in the UI have been made to accommodate this change, such as the conditional rendering of certain elements depending on whether the list title is 'Favoris' or not. Additionally, the exported default component's name from AnimeInListItem.jsx has been changed to AnimeItem.